requirements.txt
Package                    Version
-------------------------- ------------
anyascii                   0.3.2
asgiref                    3.8.1
beautifulsoup4             4.12.3
certifi                    2024.2.2
cffi                       1.16.0
chardet                    5.2.0
charset-normalizer         3.3.2
colorama                   0.4.6
crayons                    0.4.0
cryptography               42.0.5
cssselect2                 0.7.0
defusedxml                 0.7.1
Django                     5.0.6
django-allauth             0.62.1
django-bootstrap-sidebar   1.2.4
django-ckeditor            6.7.1
django-db-file-storage     0.5.6.1
django-encrypted-filefield 0.3.0
django-encrypted-files     0.0.10
django-encryption          0.3.29
django-file-form           3.6.0
django-filebrowser         4.0.3
django-filefield-encrypted 0.1.1
django-filer               3.1.1
django-filter              24.2
django-grappelli           3.0.9
django-js-asset            2.2.0
django-modelcluster        6.3
django-permissionedforms   0.1
django-polymorphic         3.1.0
django-taggit              5.0.1
django-treebeard           4.7.1
djangorestframework        3.15.1
draftjs-exporter           5.0.0
easy-thumbnails            2.8.5
et-xmlfile                 1.1.0
filetype                   1.2.0
idna                       3.7
l18n                       2021.3
laces                      0.1.1
lxml                       5.2.1
msoffcrypto-tool           5.4.0
mypy                       1.10.0
mypy-extensions            1.0.0
numpy                      1.26.4
oauthlib                   3.2.2
olefile                    0.47
opencv-python              4.9.0.80
openpyxl                   3.1.2
packaging                  24.1
piexif                     1.1.3
pillow                     10.3.0
pillow_heif                0.16.0
pip                        24.1.2
psycopg2                   2.9.9
pyAesCrypt                 6.1.1
pyasn1                     0.6.0
pycparser                  2.22
PyJWT                      2.8.0
PyMuPDF                    1.24.4
PyMuPDFb                   1.24.3
pyotp                      2.9.0
PyPDF4                     1.27.0
pypng                      0.20220715.0
pytesseract                0.3.10
python-dotenv              1.0.1
python-magic               0.4.27
python3-openid             3.2.0
pytz                       2024.1
qrcode                     7.4.2
reportlab                  4.2.0
requests                   2.31.0
requests-oauthlib          2.0.0
rsa                        4.9
setuptools                 69.5.1
six                        1.16.0
social-auth-app-django     5.4.1
social-auth-core           4.5.4
soupsieve                  2.5
sqlparse                   0.5.0
stegano                    0.11.3
svglib                     1.5.1
telepath                   0.3.1
tinycss2                   1.3.0
typing_extensions          4.11.0
tzdata                     2024.1
urllib3                    2.2.1
wagtail                    6.1
webencodings               0.5.1
Willow                     1.8.0
